refactored update endpoints
